* Add Eclipse Favorites - for static import of 'org.testng.Assert.*' * won't auto add again if people remove it
