+++ testng-eclipse-plugin/src/main/org/testng/eclipse/TestNGPluginConstants.java
@@ -32,6 +32,7 @@ public abstract class TestNGPluginConstants {
   public static final String S_PRE_DEFINED_LISTENERS = "preDefinedListeners";
   public static final String S_SHOW_VIEW_WHEN_TESTS_COMPLETE = "showViewWhenTestComplete";
   public static final String S_VIEW_TITLE_SHOW_CASE_NAME = "showCaseNameOnViewTitle";
+  public static final String S_APPEND_FAVORITE_STATIC_IMPORT = "appendFavoriteStaticImport";
 
   private TestNGPluginConstants() {}
 }
